formatting, refactoring, and fix failing test for interactive login
