+++ src/main/kotlin/no/nav/security/mock/oauth2/OAuth2Config.kt
@@ -1,6 +1,7 @@
 package no.nav.security.mock.oauth2
 
-import no.nav.security.mock.oauth2.callback.TokenCallback
+import no.nav.security.mock.oauth2.token.OAuth2TokenProvider
+import no.nav.security.mock.oauth2.token.TokenCallback
 
 data class OAuth2Config(
     val interactiveLogin: Boolean,
