+++ src/main/kotlin/no/nav/security/mock/oauth2/grant/ClientCredentialsGrantHandler.kt
@@ -1,10 +1,10 @@
 package no.nav.security.mock.oauth2.grant
 
 import com.nimbusds.oauth2.sdk.TokenRequest
-import no.nav.security.mock.oauth2.callback.TokenCallback
+import no.nav.security.mock.oauth2.token.TokenCallback
 import no.nav.security.mock.oauth2.extensions.expiresIn
 import no.nav.security.mock.oauth2.http.OAuth2TokenResponse
-import no.nav.security.mock.oauth2.OAuth2TokenProvider
+import no.nav.security.mock.oauth2.token.OAuth2TokenProvider
 import okhttp3.HttpUrl
 import java.util.UUID
