+++ src/main/kotlin/no/nav/security/mock/oauth2/grant/GrantHandler.kt
@@ -1,7 +1,7 @@
 package no.nav.security.mock.oauth2.grant
 
 import com.nimbusds.oauth2.sdk.TokenRequest
-import no.nav.security.mock.oauth2.callback.TokenCallback
+import no.nav.security.mock.oauth2.token.TokenCallback
 import no.nav.security.mock.oauth2.http.OAuth2TokenResponse
 import okhttp3.HttpUrl
