+++ src/test/kotlin/no/nav/security/mock/oauth2/grant/AuthorizationCodeHandlerTest.kt
@@ -12,7 +12,7 @@ import com.nimbusds.oauth2.sdk.id.ClientID
 import com.nimbusds.oauth2.sdk.id.State
 import com.nimbusds.openid.connect.sdk.AuthenticationRequest
 import com.nimbusds.openid.connect.sdk.AuthenticationSuccessResponse
-import no.nav.security.mock.oauth2.callback.DefaultTokenCallback
+import no.nav.security.mock.oauth2.token.DefaultTokenCallback
 import no.nav.security.mock.oauth2.login.Login
 import okhttp3.HttpUrl
 import okhttp3.HttpUrl.Companion.toHttpUrl
@@ -42,7 +42,6 @@ internal class AuthorizationCodeHandlerTest {
 
     @Test
     fun tokenResponse() {
-
     }
 
     @Test
@@ -63,7 +62,7 @@ internal class AuthorizationCodeHandlerTest {
         assertThatAuthResponseContainsRequiredParams(response)
 
         val tokenResponse = handler.tokenResponse(
-            tokenRequest(response.authorizationCode,"http://redirect", "openid"),
+            tokenRequest(response.authorizationCode, "http://redirect", "openid"),
             "http://myissuer".toHttpUrl(),
             DefaultTokenCallback()
         )
@@ -72,7 +71,7 @@ internal class AuthorizationCodeHandlerTest {
         assertThat(idToken.jwtClaimsSet.subject).isEqualTo("foo")
     }
 
-    private fun assertThatAuthResponseContainsRequiredParams(response: AuthenticationSuccessResponse){
+    private fun assertThatAuthResponseContainsRequiredParams(response: AuthenticationSuccessResponse) {
         assertThat(response.impliedResponseType().impliesCodeFlow()).isTrue()
         assertThat(response.impliedResponseMode()).isEqualTo(ResponseMode.QUERY)
         assertThat(response.state).isEqualTo(State("someState"))
